[Perception] add CUDA implementation of connected component generator (#3743)
* add CUDA implementation of connected componenet generator

* integrate ConnectedComponentGeneratorGPU into 'cuda_util'

+++ modules/perception/obstacle/camera/detector/yolo_camera_detector/yolo_camera_detector.h
@@ -94,9 +94,9 @@ class YoloCameraDetector : public BaseCameraDetector {
   std::shared_ptr<SyncedMemory> anchor_ = nullptr;
   int height_ = 0;
   int width_ = 0;
-  float min_2d_height_ = 0;
-  float min_3d_height_ = 0;
-  float top_k_ = 1000;
+  float min_2d_height_ = 0.0f;
+  float min_3d_height_ = 0.0f;
+  int top_k_ = 1000;
   int obj_size_ = 0;
   int output_height_ = 0;
   int output_width_ = 0;
@@ -108,9 +108,9 @@ class YoloCameraDetector : public BaseCameraDetector {
   std::vector<ObjectType> types_;
   int offset_y_ = 0;
   NMSParam nms_;
-  float inter_cls_nms_thresh_ = 1;
-  float cross_class_merge_threshold_ = 1;
-  float confidence_threshold_ = 0.1;
+  float inter_cls_nms_thresh_ = 1.0f;
+  float cross_class_merge_threshold_ = 1.0f;
+  float confidence_threshold_ = 0.1f;
   std::shared_ptr<BaseProjector> projector_;
   obstacle::yolo::YoloParam yolo_param_;
   int image_height_ = 0;
